AK: Rework error types to store kind as an enum
This commit makes windows errors store the code instead of being a
formatted string literal. This will allow comparing against the error,
and checkign what it is. The formatting is now done in the formatter,
and moved to an implementation file to avoid polluting headers with
windows.h. The kind of the error is now determined by an enum Kind which
will allow matching against the error kind.