LibCore+LibWebView: Restore was_exit_requested to EventLoop
This method was removed in 941c4df6b70229e5584e14ae5512cb9db538d30b

However, it was not exactly *unused* as the commit message would say.
This method was the only thing that allowed spin_until to exit when
the event loop was cancelled. This happens normally when IPC connections
are closed, but also when the process is killed.

The logic to properly handle process exit from event loop spins needs to
actually notify the caller that their goal condition was not met though.
That will be handled in a later commit.