IDLGenerators: Fix codegen bug with create_data_property
The create_data_property needs a variable that is generated in
generate_wrap_statement but when a item is optional it assignment
return_expression is put in a if block. I have added a trailing template
argument where you can insert code in both if else blocks,
so that the variable is accessible.